Update package versions.
This allows us to use this library with recent versions of Bucklescript
and to get rid of the dependency on rollup-watch, which depends on a
version of fsevents that doesn't work with Node 10, but which has been
incorporated into recent versions of Rollup itself.

To compile with recent Bucklescript, use true instead of Js.true_.

To compile with recent Rollup, rename the module to bs_snabbdom instead
of bs-snabbdom.